requirements.txt
Flask==2.0.3
PyMySQL==1.0.2
PyJWT==2.3.0
Flask-Redis==0.4.0
requests~=2.32.0
redis~=6.1.0
asyncio==3.4.3
fastapi==0.70.0
uvicorn==0.15.0
pytest~=8.2
pyyaml~=6.0
pyyaml-include~=1.3.1
allure-pytest~=2.13.5
pytest-json~=0.4.0
pytest-json-report~=1.5.0
pytest-xdist~=3.6.1
pytest-html~=4.1.1
retry~=0.9.2
locust~=2.37.5
bcrypt~=3.2.0
